Add datasource watch in snapshot controller as well
This is also needed for snapshot clones, guess it just merged
around the time they were introduced.
